First_app/requirements.txt
Flask==1.0.2
paho-mqtt
nano